refactor: [STEP2] KoreanEntryListTableViewDelegate 개선 및 코드 정리 출력 코드 제거, 공백 제거 TablewViewDelegate, TableViewDataSource를 채택하여 구현한 클래스 파일로 분리
